Update FM_Buffs_En.csv
I've mostly tried to avoid changing the names of effects as this is a larger change, and a lot of the time I'm not sure what you're trying to mean, or what the picture on the card is. A few like "ripple breathing" - line 304 don't make much sense, but I'm not sure what you wanted to write.